Push solving errors
